new release

+++ frontend/lib/backend/failedRequestManager.dart
@@ -73,7 +73,7 @@ _retryFailedRequestsIsolate(
             body: 'Offline Änderungen werden hochsynchronisiert...',
             category: NotificationCategory.Progress,
             notificationLayout: NotificationLayout.ProgressBar,
-            progress: (i / total * 100).round(),
+            progress: (i / total * 100).round().toDouble(),
             // locked: true,
             payload: NotificationPayload.progress(i, total)),
       );
@@ -100,7 +100,7 @@ _retryFailedRequestsIsolate(
                   'Offline Änderungen konnten NICHT(!) hochsynchronisiert werden.. probiers nochmal oder melde dich beim Support',
               category: NotificationCategory.Progress,
               notificationLayout: NotificationLayout.ProgressBar,
-              progress: (i / total * 100).round(),
+              progress: (i / total * 100).round().toDouble(),
               payload: NotificationPayload.failed(rd),
               // locked: false,
             ),
